Split out directory watch creation tests
